Betyg 4

+++ src/main/java/bo/UserHandler.java
@@ -7,18 +7,25 @@ import java.util.ArrayList;
 
 public class UserHandler {
 
-    public static boolean Login(String username, String password) {
+    public static boolean Login(String username, String password, Role role) {
         User user = UserDB.searchByUsername(username);
 
         if (user == null){
             return false;
         }
 
-        return user.getPassword().equals(password);
+        System.out.println(user);
+        return user.getPassword().equals(password) && user.getRole().equals(role);
     }
 
     public static ArrayList<Item> getMyItems(String username) {
-        ArrayList<Item> myItems = ItemDB.searchItemByUsername(username);
+        User user = UserDB.searchByUsername(username);
+
+        if (user == null){
+            return new ArrayList<>();
+        }
+
+        ArrayList<Item> myItems = ItemDB.searchItemByUsername(username, user.getId());
 
         return myItems;
     }
@@ -30,10 +37,23 @@ public class UserHandler {
     }
 
     public static boolean addItem(int userId, int itemId) {
+        User user = UserDB.searchById(userId);
+        Item item = ItemDB.searchById(itemId);
+
+        // Kontrollera om användaren eller varan inte finns
+        if (user == null || item == null) {
+            return false;
+        }
+
+        if (item.getBalance() < 1){
+            return false;
+        }
+
+
         if (UserDB.addItem(userId, itemId)){
-            User user = UserDB.searchById(userId);
-            Item item = ItemDB.searchById(itemId);
             user.addItem(item);
+            item.setBalance(item.getBalance() - 1);
+            ItemDB.updateBalance(itemId, item.getBalance());
             return true;
         }
 
@@ -41,10 +61,18 @@ public class UserHandler {
     }
 
     public static boolean removeItem(int userId, int itemId) {
+        User user = UserDB.searchById(userId);
+        Item item = ItemDB.searchById(itemId);
+
+        // Kontrollera om användaren eller varan inte finns
+        if (user == null || item == null) {
+            return false;
+        }
+
         if (UserDB.removeItem(userId, itemId)){
-            User user = UserDB.searchById(userId);
-            Item item = ItemDB.searchById(itemId);
             user.removeItem(item);
+            item.setBalance(item.getBalance() + 1);
+            ItemDB.updateBalance(itemId, item.getBalance());
             return true;
         }
